requirements.txt
# Core libraries
pandas==1.5.3
pytest==8.3.4
pyyaml==6.0.2
